+++ src/index.ts
@@ -1,20 +1,14 @@
 #!/usr/bin/env node
+import { build } from './build'
 import getOptions from './getOptions'
 
-const options = getOptions()
-
 const main = async () => {
+  const options = getOptions()
   if (options.version) {
     const { displayVersion } = await import('./displayVersion')
     return displayVersion()
   }
-  const { watchAndBuild } = await import('./watchAndBuild')
-  options.watch
-    ? (await import('chokidar'))
-        .watch(options.watch)
-        .on('ready', async () => watchAndBuild())
-        .on('change', async () => watchAndBuild())
-    : watchAndBuild()
+  await build(options)
 }
 
 main()